Split data file and update for 4.0
Splits locations into separate files so it's easier to understand/update. In order to do this, I had to structure the data differently. Now there's a 'Section' that contains 'Categories' which contain 'Locations'.

3.0 Locations not shown by default

Updated region markers to allow deleting (right-click on a marker deletes them) so it's easier to mark new regions.

Updates handlebars to latest (v1 to v4)
Updates colors to be similar to 4.0
Fixes some small alignment issues